feat: create whkd_core crate
This commit creates a `whkd_core` crate which contains all data structs
required by `whkd`. This will allow 3rd party apps to access this data
to interface with whkd if needed.
