+++ .gitignore
@@ -1,2 +1,2 @@
-/target
+target/
 .idea
